feature Add the configuration status and button to configure
Change the Login Status to display the text if the extension is configured or not.

Add a button to open the chrome extension options page to make it easy to configure the extension if not configured.
Update the code to check anytime the tab state has been updated in order to check if the extension is configured.